cxl/events: Promote CXL event structures to a core header
commit 26a1a86dd093a10d0653429bf013dae6e95dccbf upstream.

UEFI code can process CXL events through CPER records.  Those records
use almost the same format as the CXL events.

Lift the CXL event structures to a core header to be shared in later
patches.

[jic123: drop "CXL rev 3.0" mention]

Link: https://lore.kernel.org/r/20231220-cxl-cper-v5-2-1bb8a4ca2c7a@intel.com
[djbw: add F: entry to maintainers for include/linux/cxl-event.h]
Acked-by: Ard Biesheuvel <ardb@kernel.org>
